CMakeList.txt adjustements VST3 Works!
Made some adjustments to the CMakeList.txt file so that the compile would function for VST3.

Changed COPY_PLUGIN_AFTER_BUILD to FALSE as powershell wasn't run as admin and didn't have permissions to  install.  Which is fine for now for dev work.

Added in my JuceIcnlude.h again

Moved the if(WIn32) above the FetchContent. Don't think this mattered but I was desperate for a fix. :)